+++ right/src/keyboard_layout.h
@@ -3,6 +3,7 @@
 
 #include <stdint.h>
 #include "lufa/HIDClassCommon.h"
+#include "usb_composite_device.h"
 
 /**
  * Keyboard layout is a 2D array of scan codes.
@@ -35,53 +36,6 @@
 #define MODIFIER_MOD_PRESSED 1
 #define MODIFIER_FN_PRESSED 2
 
-extern uint8_t keyMasks[LAYOUT_KEY_COUNT];
-
-
-static inline __attribute__((always_inline)) uint8_t getKeycode(KEYBOARD_LAYOUT(layout), uint8_t keyId, uint8_t modifierState)
-{
-	if (keyId<LAYOUT_KEY_COUNT) {
-		if (keyMasks[keyId]!=0 && keyMasks[keyId]!=modifierState){
-			//Mask out key presses after releasing modifier keys
-			return 0;
-		}
-
-		uint8_t k = layout[keyId][modifierState];
-		keyMasks[keyId] = modifierState;
-
-		if (k==0) {
-			k = layout[keyId][0];
-		}
-
-		return k;
-	} else {
-		return 0;
-	}
-}
-
-static inline __attribute__((always_inline)) uint8_t getModifierState(const uint8_t *leftKeyStates, const uint8_t *rightKeyStates){
-	uint8_t mod = 0;
-	if (leftKeyStates[KEYID_LEFT_MOD]) {
-		mod |= MODIFIER_MOD_PRESSED;
-	}
-	if (leftKeyStates[KEYID_LEFT_FN] | rightKeyStates[KEYID_RIGHT_FN]) {
-		mod |= MODIFIER_FN_PRESSED;
-	}
-
-	return mod;
-}
-
-static inline __attribute__((always_inline)) void clearKeymasks(const uint8_t *leftKeyStates, const uint8_t *rightKeyStates){
-	int i;
-	for (i=0; i<KEY_STATE_COUNT; ++i){
-		if (rightKeyStates[i]==0){
-			keyMasks[i] = 0;
-		}
-
-		if (leftKeyStates[i]==0){
-			keyMasks[LAYOUT_LEFT_OFFSET+i] = 0;
-		}
-	}
-}
+void fillKeyboardReport(usb_keyboard_report_t *report, const uint8_t *leftKeyStates, const uint8_t *rightKeyStates, KEYBOARD_LAYOUT(layout));
 
 #endif
